requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
cookies==2.2.1
idna==3.10
instaloader==4.14.2
requests==2.32.4
tqdm==4.67.1
urllib3==2.5.0
instaloader==4.14.2
pandas==2.2.2
python-dotenv==1.0.1